+++ RecognizerBrain/Domain/Models/AddRecognitionNodeModel.cs
@@ -3,7 +3,7 @@ namespace Domain.Models
     public record AddRecognitionNodeModel
     {
         public long TrackId { get; init; }
-        public long IdentificationHash { get; init; }
+        public uint IdentificationHash { get; init; }
         public int Duration { get; init; }
     }
 }
