+++ RecognizerBrain/Domain/Models/FindRecognitionNodeModel.cs
@@ -2,7 +2,7 @@ namespace Domain.Models
 {
     public record class FindRecognitionNodeModel
     {
-        public long IdentificationHash { get; init; }
+        public uint IdentificationHash { get; init; }
         public int Duration { get; init; } 
     }
 }
